Begin fixing from compilation side
Removed floating image file, added .gitignore to prevent pycache being
tracked. Began process of creating documentation on this project.

Currently docs/modules will be the details to use; as things stabilise
this will be turned into docstrings and docs/howto should replace it as
the canonical source. The goal is that any part of the software can be
replaced with another implementation which inherits the basics, and will
be fine.